requirements.txt
certifi
chardet
idna
numpy
python-dotenv
requests
urllib3
